+++ tests/test_balances.py
@@ -6,7 +6,8 @@ import pytest
 from web3 import Web3, EthereumTesterProvider
 from web3.contract import Contract
 
-from eth_hentai.balances import fetch_erc20_balances, fetch_erc20_balances_decimal
+from eth_hentai.balances import fetch_erc20_balances_by_transfer_event, fetch_erc20_balances_decimal_by_transfer_event, \
+    fetch_erc20_balances_by_token_list
 from eth_hentai.token import create_token
 
 
@@ -70,7 +71,7 @@ def test_portfolio_current(web3: Web3, deployer: str, user_1: str, usdc: Contrac
     # Load up the user with some tokens
     usdc.functions.transfer(user_1, 500).transact({"from": deployer})
     aave.functions.transfer(user_1, 200).transact({"from": deployer})
-    balances = fetch_erc20_balances(web3, user_1)
+    balances = fetch_erc20_balances_by_transfer_event(web3, user_1)
     assert balances[usdc.address] == 500
     assert balances[aave.address] == 200
 
@@ -87,11 +88,11 @@ def test_portfolio_past(web3: Web3, deployer: str, user_1: str, usdc: Contract, 
     # Top up AAVE which won't show up in the analysis
     aave.functions.transfer(user_1, 333).transact({"from": deployer})
 
-    balances = fetch_erc20_balances(web3, user_1, last_block_num=threshold_block)
+    balances = fetch_erc20_balances_by_transfer_event(web3, user_1, last_block_num=threshold_block)
     assert balances[usdc.address] == 500
     assert balances[aave.address] == 200
 
-    balances = fetch_erc20_balances(web3, user_1)
+    balances = fetch_erc20_balances_by_transfer_event(web3, user_1)
     assert balances[aave.address] == 533
 
 
@@ -101,7 +102,7 @@ def test_portfolio_decimals(web3: Web3, deployer: str, user_1: str, usdc: Contra
     # Load up the user with some tokens
     usdc.functions.transfer(user_1, 500).transact({"from": deployer})
     aave.functions.transfer(user_1, 200).transact({"from": deployer})
-    balances = fetch_erc20_balances_decimal(web3, user_1)
+    balances = fetch_erc20_balances_decimal_by_transfer_event(web3, user_1)
     assert balances[usdc.address].value == Decimal('0.0005')
     assert balances[usdc.address].decimals == 6
     assert balances[aave.address].value == Decimal(200) / Decimal(10**18)
@@ -112,7 +113,7 @@ def test_portfolio_two_transactions(web3: Web3, deployer: str, user_1: str, usdc
     """Get the balance after two top up transactions."""
     usdc.functions.transfer(user_1, 500).transact({"from": deployer})
     usdc.functions.transfer(user_1, 300).transact({"from": deployer})
-    balances = fetch_erc20_balances(web3, user_1)
+    balances = fetch_erc20_balances_by_transfer_event(web3, user_1)
     assert balances[usdc.address] == 800
 
 
@@ -120,5 +121,17 @@ def test_portfolio_debit_transactions(web3: Web3, deployer: str, user_1: str, us
     """Get the balance after debit  transactions."""
     usdc.functions.transfer(user_1, 500).transact({"from": deployer})
     usdc.functions.transfer(deployer, 300).transact({"from": user_1})
-    balances = fetch_erc20_balances(web3, user_1)
+    balances = fetch_erc20_balances_by_transfer_event(web3, user_1)
     assert balances[usdc.address] == 200
+
+
+def test_portfolio_token_list(web3: Web3, deployer: str, user_1: str, usdc: Contract, aave: Contract):
+    """Analyse current holdings by a token list."""
+    # Create a set of tokens
+    tokens = {aave.address, usdc.address}
+    # Load up the user with some tokens
+    usdc.functions.transfer(user_1, 500).transact({"from": deployer})
+    aave.functions.transfer(user_1, 200).transact({"from": deployer})
+    balances = fetch_erc20_balances_by_token_list(web3, user_1, tokens)
+    assert balances[usdc.address] == 500
+    assert balances[aave.address] == 200